Move `Declare*` node to `flow.js`
