+++ src/language-js/printer-estree.js
@@ -72,7 +72,6 @@ const {
   printOptionalToken,
   printBindExpressionCallee,
   printTypeScriptModifiers,
-  printFlowDeclaration,
   adjustClause,
 } = require("./print/misc");
 const {
@@ -89,6 +88,7 @@ const { printObject } = require("./print/object");
 const {
   printTypeAnnotation,
   shouldHugType,
+  printOpaqueType,
 } = require("./print/type-annotation");
 const {
   printClass,
@@ -491,10 +491,8 @@ function printPathNoParens(path, options, print, args) {
       return concat(["export = ", path.call(print, "expression"), semi]);
     case "ExportDefaultDeclaration":
     case "ExportNamedDeclaration":
-    case "DeclareExportDeclaration":
       return printExportDeclaration(path, options, print);
     case "ExportAllDeclaration":
-    case "DeclareExportAllDeclaration":
       return printExportAllDeclaration(path, options, print);
     case "ImportDeclaration":
       return printImportDeclaration(path, options, print);
@@ -1158,66 +1156,8 @@ function printPathNoParens(path, options, print, args) {
         printFunctionDeclaration(path, print, options),
         semi,
       ]);
-    case "DeclareFunction":
-      return printFlowDeclaration(
-        path,
-        concat([
-          "function ",
-          path.call(print, "id"),
-          n.predicate ? " " : "",
-          path.call(print, "predicate"),
-          semi,
-        ])
-      );
-    case "DeclareModule":
-      return printFlowDeclaration(
-        path,
-        concat([
-          "module ",
-          path.call(print, "id"),
-          " ",
-          path.call(print, "body"),
-        ])
-      );
-    case "DeclareModuleExports":
-      return printFlowDeclaration(
-        path,
-        concat([
-          "module.exports",
-          ": ",
-          path.call(print, "typeAnnotation"),
-          semi,
-        ])
-      );
-    case "DeclareVariable":
-      return printFlowDeclaration(
-        path,
-        concat(["var ", path.call(print, "id"), semi])
-      );
-    case "DeclareOpaqueType":
-    case "OpaqueType": {
-      parts.push(
-        "opaque type ",
-        path.call(print, "id"),
-        path.call(print, "typeParameters")
-      );
-
-      if (n.supertype) {
-        parts.push(": ", path.call(print, "supertype"));
-      }
-
-      if (n.impltype) {
-        parts.push(" = ", path.call(print, "impltype"));
-      }
-
-      parts.push(semi);
-
-      if (n.type === "DeclareOpaqueType") {
-        return printFlowDeclaration(path, concat(parts));
-      }
-
-      return concat(parts);
-    }
+    case "OpaqueType":
+      return printOpaqueType(path, print, options);
 
     case "EnumDeclaration":
       return concat([
@@ -1390,7 +1330,6 @@ function printPathNoParens(path, options, print, args) {
       ]);
     }
 
-    case "DeclareInterface":
     case "InterfaceDeclaration":
     case "InterfaceTypeAnnotation":
     case "TSInterfaceDeclaration":
@@ -1575,31 +1514,8 @@ function printPathNoParens(path, options, print, args) {
         return printNumber(n.extra.raw);
       }
       return printNumber(n.raw);
-
-    case "DeclareTypeAlias":
-    case "TypeAlias": {
-      if (n.type === "DeclareTypeAlias" || n.declare) {
-        parts.push("declare ");
-      }
-
-      const printed = printAssignmentRight(
-        n.id,
-        n.right,
-        path.call(print, "right"),
-        options
-      );
-
-      parts.push(
-        "type ",
-        path.call(print, "id"),
-        path.call(print, "typeParameters"),
-        " =",
-        printed,
-        semi
-      );
-
-      return group(concat(parts));
-    }
+    case "TypeAlias":
+      return printTypeAlias(path, options, print);
     case "TypeCastExpression": {
       return concat([
         "(",
